+++ world/src/repositories/creature_static_data.rs
@@ -70,7 +70,7 @@ impl CreatureBaseAttributesPerLevelDbRecord {
         ((self.health[expansion] as f32 * multiplier).ceil() as u32).max(1)
     }
 
-    pub fn damage(&self, expansion: usize, multiplier: f32) -> u32 {
-        ((self.damage[expansion] as f32 * multiplier).ceil() as u32).max(1)
+    pub fn damage(&self, expansion: usize, multiplier: f32) -> f32 {
+        ((self.damage[expansion] as f32 * multiplier).ceil()).max(1.)
     }
 }
